Squashed commit of the following:
commit d2884be5fef4962377550feba0ea5ff5984d3c4d
Author: Aleksthor <athorset98@gmail.com>
Date:   Mon Jan 30 20:42:04 2023 +0100

    Barebones Jeger AI

+++ UNITY/MooseOrLose/Assets/Scripts/BehaviorTree/Tasks/Elg/ElgRunAway.cs
@@ -21,7 +21,7 @@ public class ElgRunAway : Node
         {
             return NodeState.FAILURE;
         }
-        float speed = runSpeed * ((float)TimeManager.instance.startPlaySpeed / (float)TimeManager.instance.playSpeed);
+        float speed = runSpeed * (TimeManager.instance.startPlaySpeed / TimeManager.instance.playSpeed);
         mAgent.speed = speed;
         Transform Danger = (Transform)parent.GetData("Danger");
         Vector3 direction = mTransform.position - Danger.position;
